fixed some x64 compiler errors/warnings we still need a GLEW x64 lib
git-svn-id: https://svn.code.sf.net/p/vbam/code/trunk@219 a31d4220-a93d-0410-bf67-fe4944624d44

+++ src/win32/SelectPlugin.h
@@ -20,7 +20,7 @@ class SelectPlugin : public CDialog
 // Construction
 public:
 	SelectPlugin(CWnd* pParent = NULL);   // standard constructor
-	int EnumPlugins();
+	size_t EnumPlugins();
 	bool GetPluginDesc(const char *sRpi, PluginDesc *pDesc);
 
 // Dialog Data
